# backend/docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Rug Radar Token Module API",
    "description": "API untuk mendeteksi token, membaca bytecode, dan mengelola penilaian risiko Rug Radar.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
    "/tokens": {
      "get": {
        "summary": "List semua token",
        "description": "Mengembalikan daftar token dengan paginasi dan filter pencarian.",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Halaman data"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20
            },
            "description": "Jumlah item per halaman"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Pencarian alamat token (case-insensitive)"
          }
        ],
        "responses": {
          "200": {
            "description": "List token berhasil diambil",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Token"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PaginationMeta"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tokens/{address}": {
      "get": {
        "summary": "Detail token",
        "description": "Mengembalikan detail token beserta assessment risiko terbarunya.",
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^0x[0-9a-fA-F]{40}$"
            },
            "description": "Alamat Ethereum token"
          }
        ],
        "responses": {
          "200": {
            "description": "Token ditemukan",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/TokenDetail"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Format alamat tidak valid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Token tidak ditemukan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/tokens/{address}/assessments": {
      "get": {
        "summary": "Riwayat risk assessments token",
        "description": "Mengembalikan daftar semua risk assessments untuk token tertentu.",
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Alamat Ethereum token"
          }
        ],
        "responses": {
          "200": {
            "description": "Riwayat assessment berhasil diambil",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/RiskAssessment"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Format alamat tidak valid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Token tidak ditemukan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Token": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "address": {
            "type": "string"
          },
          "chain_id": {
            "type": "integer"
          },
          "deployer": {
            "type": "string"
          },
          "deployed_at": {
            "type": "string",
            "format": "date-time"
          },
          "has_unlimited_mint": {
            "type": "boolean",
            "nullable": true
          },
          "has_blacklist": {
            "type": "boolean",
            "nullable": true
          },
          "has_tax": {
            "type": "boolean",
            "nullable": true
          },
          "liquidity_locked": {
            "type": "boolean",
            "nullable": true
          },
          "top_holder_concentration": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "RiskAssessment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "token_id": {
            "type": "string",
            "format": "uuid"
          },
          "probability": {
            "type": "number",
            "format": "double"
          },
          "reasoning": {
            "type": "string"
          },
          "confidence": {
            "type": "number",
            "format": "double"
          },
          "llm_model": {
            "type": "string"
          },
          "assessed_at": {
            "type": "string",
            "format": "date-time"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "TokenDetail": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Token"
          },
          {
            "type": "object",
            "properties": {
              "latest_assessment": {
                "$ref": "#/components/schemas/RiskAssessment",
                "nullable": true
              }
            }
          }
        ]
      },
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "error"
          },
          "message": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        }
      }
    }
  }
}